requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
asgiref==3.5.1
    # via
    #   -r ./requirements.txt
    #   django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via pytest
backcall==0.2.0
    # via ipython
black==22.3.0
    # via -r requirements_dev.in
brotli==1.0.9
    # via geventhttpclient
certifi==2021.10.8
    # via
    #   -r ./requirements.txt
    #   geventhttpclient
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.12
    # via
    #   -r ./requirements.txt
    #   requests
click==8.1.3
    # via
    #   black
    #   flask
    #   pip-tools
configargparse==1.5.3
    # via locust
coreapi==2.3.3
    # via
    #   -r ./requirements.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-xml
django==4.0.4
    # via
    #   -r ./requirements.txt
    #   django-datetime-utc
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-yasg
    #   model-bakery
django-datetime-utc==1.0.3
    # via -r ./requirements.txt
django-debug-toolbar==3.4.0
    # via -r ./requirements.txt
django-extensions==3.1.5
    # via -r ./requirements.txt
django-filter==21.1
    # via -r ./requirements.txt
djangorestframework==3.13.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
    #   djangorestframework-gis
    #   drf-amsterdam
    #   drf-extensions
    #   drf-yasg
djangorestframework-csv==2.1.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
djangorestframework-gis==1.0
    # via -r ./requirements.txt
djangorestframework-xml==2.0.0
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-amsterdam==0.4.1
    # via -r ./requirements.txt
drf-extensions==0.7.1
    # via
    #   -r ./requirements.txt
    #   drf-amsterdam
drf-yasg==1.20.0
    # via -r ./requirements.txt
executing==0.8.3
    # via stack-data
factory-boy==3.2.1
    # via -r requirements_dev.in
faker==13.11.0
    # via factory-boy
flask==2.1.2
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==3.0.10
    # via locust
gevent==21.12.0
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==1.5.3
    # via locust
greenlet==1.1.2
    # via gevent
idna==3.3
    # via
    #   -r ./requirements.txt
    #   requests
inflection==0.5.1
    # via
    #   -r ./requirements.txt
    #   drf-yasg
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements_dev.in
ipython==8.3.0
    # via ipdb
isort==5.10.1
    # via -r requirements_dev.in
itsdangerous==2.1.2
    # via flask
itypes==1.2.0
    # via
    #   -r ./requirements.txt
    #   coreapi
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   -r ./requirements.txt
    #   coreschema
    #   flask
    #   locust
locust==2.8.6
    # via -r requirements_dev.in
markupsafe==2.1.1
    # via
    #   -r ./requirements.txt
    #   jinja2
matplotlib-inline==0.1.3
    # via ipython
model-bakery==1.5.0
    # via -r requirements_dev.in
msgpack==1.0.3
    # via locust
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   -r ./requirements.txt
    #   drf-yasg
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.6.0
    # via -r requirements_dev.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.29
    # via ipython
psutil==5.9.0
    # via locust
psycopg2-binary==2.9.3
    # via -r ./requirements.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pygments==2.12.0
    # via ipython
pyparsing==3.0.9
    # via
    #   -r ./requirements.txt
    #   packaging
pytest==7.1.2
    # via
    #   -r requirements_dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements_dev.in
python-dateutil==2.8.2
    # via
    #   -r ./requirements.txt
    #   django-datetime-utc
    #   faker
    #   time-machine
pytz==2022.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework
pyzmq==22.3.0
    # via locust
requests==2.27.1
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   locust
roundrobin==0.0.2
    # via locust
ruamel.yaml==0.17.21
    # via
    #   -r ./requirements.txt
    #   drf-yasg
ruamel.yaml.clib==0.2.6
    # via
    #   -r ./requirements.txt
    #   ruamel.yaml
sentry-sdk==1.5.12
    # via -r ./requirements.txt
six==1.16.0
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
    #   flask-cors
    #   geventhttpclient
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   -r ./requirements.txt
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
time-machine==2.7.0
    # via -r requirements_dev.in
toml==0.10.2
    # via ipdb
tomli==2.0.1
    # via
    #   black
    #   pep517
    #   pytest
traitlets==5.2.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.2.0
    # via locust
unicodecsv==0.14.1
    # via
    #   -r ./requirements.txt
    #   djangorestframework-csv
uritemplate==4.1.1
    # via
    #   -r ./requirements.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via
    #   -r ./requirements.txt
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.1.2
    # via
    #   flask
    #   locust
wheel==0.37.1
    # via pip-tools
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
